Save inputs to provide defaults for next session.
The inputs provided for interface, victim ip, and gateway ip are saved
in a file within the scripts directory (.last_session). On startup, this
file is read, and if found, used to provide defaults for the
inputs (shown in parentheses). Providing an empty input will make the
script use the default it read from the .last_session file.